refactor: move order status constants
