+++ app/domain/order/constants.ts
@@ -1,0 +1,8 @@
+import type { OrderStatus } from "./entities/order"
+
+export const ORDER_STATUSES: OrderStatus[] = [
+  "pending",
+  "processing",
+  "completed",
+  "cancelled",
+]
