+++ app/domain/order/entities/order.ts
@@ -1,16 +1,8 @@
-/**
- * 注文ステータスの定義
- *
- * システム全体で使用するステータス値の正規定義
- */
-export const ORDER_STATUSES = [
-  "pending",
-  "processing",
-  "completed",
-  "cancelled",
-] as const
-
-export type OrderStatus = (typeof ORDER_STATUSES)[number]
+export type OrderStatus =
+  | "pending"
+  | "processing"
+  | "completed"
+  | "cancelled"
 
 type Order = {
   id: number
